Moved construction of the incremental `Compiler` implementation for languages that use the C preprocessor out of the task types and into the `NativeCompilerFactory` service, to live with the other pieces of incremental compilation.
This change means that the tool chain is completely responsible for incremental compilation supported for a given language + tool, rather than both the task and tool chain providing parts of this support. The tool chain can delegate to `NativeCompilerFactory` to easily add support for incremental compilation for C processor based languages.

This change adds a dependency cycle between the platformNative and languageNative projects, which will be removed in a subsequent change.