Add a `BuildOperationExecutor` backed parallel zip creator
This commit replaces the use of the Commons Compress parallel zip creator with an alternate version which
uses our `BuildOperationExecutor`. There are several differences in the implementation:

- we get `ScatterZipOutputStream` using a linked blocking queue, instead of using a `ThreadLocal`, avoiding potential memory leaks
- compression is executed on Gradle-managed threads, ensuring that we never exhaust resources when running concurrently with other tasks

We also fall back to the old mechanism as soon as _reproducible order_ is required (this might be possible to generate a reproducible
output with parallel compression too, but it requires more work).