Record for each source folder which Eclipse scope it is used by
Eclipse source folders each have a scope assigned to them. This
is currently the name of the SourceSet, but might evolve later
to be the name of the component instead. The exact value is an
implementation detail.

However, we also need to track all the scopes that *use* that
source folder, so its output can be included when running
applications and tests in Eclipse. The output folder is not
explicitly modeled in the .classpath file, but instead implicitly
belongs to the source folder.